+++ src/connectors/data_lake/delta.rs
@@ -4,7 +4,6 @@ use std::collections::{HashMap, HashSet, VecDeque};
 use std::fmt;
 use std::fs::File;
 use std::io::{Seek, SeekFrom, Write};
-use std::path::Path;
 use std::sync::Arc;
 use std::thread::sleep;
 use std::time::{Duration, Instant};
@@ -17,12 +16,12 @@ use deltalake::datafusion::prelude::Expr;
 use deltalake::datafusion::scalar::ScalarValue;
 use deltalake::kernel::Action as DeltaLakeAction;
 use deltalake::kernel::ArrayType as DeltaTableArrayType;
+use deltalake::kernel::CommitInfo as DeltaTableCommitInfo;
 use deltalake::kernel::DataType as DeltaTableKernelType;
 use deltalake::kernel::PrimitiveType as DeltaTablePrimitiveType;
 use deltalake::kernel::StructField as DeltaTableStructField;
 use deltalake::kernel::StructType as DeltaTableStructType;
 use deltalake::operations::create::CreateBuilder as DeltaTableCreateBuilder;
-use deltalake::parquet::file::reader::FileReader as DeltaLakeParquetFileReader;
 use deltalake::parquet::record::reader::RowIter as ParquetRowIterator;
 use deltalake::parquet::record::Row as ParquetRow;
 use deltalake::protocol::SaveMode as DeltaTableSaveMode;
@@ -37,11 +36,13 @@ use tokio::runtime::Runtime as TokioRuntime;
 
 use super::{
     columns_into_pathway_values, parquet_row_into_values_map, LakeBatchWriter, LakeWriterSettings,
-    MetadataPerColumn, PATHWAY_COLUMN_META_FIELD, SPECIAL_OUTPUT_FIELDS,
+    MetadataPerColumn, PATHWAY_COLUMN_META_FIELD, SPECIAL_FIELD_TIME, SPECIAL_OUTPUT_FIELDS,
 };
 use crate::async_runtime::create_async_tokio_runtime;
 use crate::connectors::data_format::parse_bool_advanced;
+use crate::connectors::data_lake::ArrowDataType;
 use crate::connectors::data_storage::{ConnectorMode, ValuesMap};
+use crate::connectors::metadata::ParquetMetadata;
 use crate::connectors::scanner::S3Scanner;
 use crate::connectors::{
     DataEventType, OffsetKey, OffsetValue, ReadError, ReadResult, Reader, ReaderContext,
@@ -386,12 +387,33 @@ impl ObjectDownloader {
 pub struct DeltaReaderAction {
     action_type: DataEventType,
     path: String,
+    is_last_in_version: bool,
 }
 
 impl DeltaReaderAction {
     pub fn new(action_type: DataEventType, path: String) -> Self {
-        Self { action_type, path }
+        Self {
+            action_type,
+            path,
+            is_last_in_version: false,
+        }
     }
+
+    pub fn set_last_in_version(&mut self) {
+        self.is_last_in_version = true;
+    }
+}
+
+#[derive(Debug)]
+enum ParquetReaderOutcome {
+    SourceEvent(ReadResult),
+    Row(ParquetRow),
+}
+
+#[derive(Debug)]
+enum BackfillingEntry {
+    SourceEvent(ReadResult),
+    Entry(ValuesMap),
 }
 
 #[allow(clippy::module_name_repetitions)]
@@ -404,11 +426,10 @@ pub struct DeltaTableReader {
 
     reader: Option<ParquetRowIterator<'static>>,
     current_version: i64,
-    last_fully_read_version: Option<i64>,
     rows_read_within_version: i64,
     parquet_files_queue: VecDeque<DeltaReaderAction>,
-    current_event_type: DataEventType,
-    backfilling_entries_queue: VecDeque<ValuesMap>,
+    current_action: Option<DeltaReaderAction>,
+    backfilling_entries_queue: VecDeque<BackfillingEntry>,
 }
 
 const APPEND_ONLY_PROPERTY_NAME: &str = "delta.appendOnly";
@@ -422,7 +443,7 @@ impl DeltaTableReader {
         path: &str,
         object_downloader: ObjectDownloader,
         storage_options: HashMap<String, String>,
-        column_types: HashMap<String, Type>,
+        mut column_types: HashMap<String, Type>,
         streaming_mode: ConnectorMode,
         start_from_timestamp_ms: Option<i64>,
         has_primary_key: bool,
@@ -444,7 +465,13 @@ impl DeltaTableReader {
             return Err(ReadError::PrimaryKeyRequired);
         }
         let mut current_version = table.version();
-        let mut parquet_files_queue = Self::get_reader_actions(&table, path)?;
+
+        let mut parquet_files_queue = {
+            let history = runtime.block_on(async {
+                Ok::<Vec<DeltaTableCommitInfo>, ReadError>(table.history(None).await?)
+            })?;
+            Self::get_reader_actions(&table, path, history)?
+        };
         let mut backfilling_entries_queue = VecDeque::new();
 
         if let Some(start_from_timestamp_ms) = start_from_timestamp_ms {
@@ -488,7 +515,7 @@ impl DeltaTableReader {
                 &runtime,
                 table.clone(),
                 backfilling_thresholds,
-                &column_types,
+                &mut column_types,
             )?;
         }
 
@@ -500,21 +527,33 @@ impl DeltaTableReader {
 
             current_version,
             object_downloader,
-            last_fully_read_version: None,
             reader: None,
             backfilling_entries_queue,
             parquet_files_queue,
             rows_read_within_version: 0,
-            current_event_type: DataEventType::Insert,
+            current_action: None,
         })
+    }
+
+    fn record_batch_has_pathway_fields(batch: &ArrowRecordBatch) -> bool {
+        for (field, _) in SPECIAL_OUTPUT_FIELDS {
+            if let Some(time_column) = batch.column_by_name(field) {
+                if *time_column.data_type() != ArrowDataType::Int64 {
+                    return false;
+                }
+            } else {
+                return false;
+            }
+        }
+        true
     }
 
     fn create_backfilling_files_queue(
         runtime: &TokioRuntime,
         table: DeltaTable,
         backfilling_thresholds: Vec<BackfillingThreshold>,
-        column_types: &HashMap<String, Type>,
-    ) -> Result<VecDeque<ValuesMap>, ReadError> {
+        column_types: &mut HashMap<String, Type>,
+    ) -> Result<VecDeque<BackfillingEntry>, ReadError> {
         let backfilling_started_at = Instant::now();
         let ctx = DeltaSessionContext::new();
         ctx.register_table("table", Arc::new(table))?;
@@ -535,24 +574,88 @@ impl DeltaTableReader {
                 ),
             };
         }
-        let mut backfilling_entries_queue = VecDeque::new();
+
+        let has_pathway_meta_column = column_types.get(SPECIAL_FIELD_TIME).is_some();
+        let mut pathway_meta_column_added = false;
+
+        let mut backfilling_entries = Vec::new();
         runtime.block_on(async {
             let results = df.collect().await?;
+            let mut is_first_entry = true;
             for entry in results {
+                if is_first_entry
+                    && !has_pathway_meta_column
+                    && Self::record_batch_has_pathway_fields(&entry)
+                {
+                    // We're dealing with the output of Pathway process. It means that it also needs to be
+                    // processed atomically with respect to times.
+                    // To accomplish that, we temporarily add Pathway's meta field "time", use it in the
+                    // batch splits, and disregard later on.
+                    column_types.insert(SPECIAL_FIELD_TIME.to_string(), Type::Int);
+                    pathway_meta_column_added = true;
+                }
                 let value_maps = columns_into_pathway_values(&entry, column_types);
                 for value_map in value_maps {
-                    backfilling_entries_queue.push_back(value_map);
+                    backfilling_entries.push(value_map);
                 }
+                is_first_entry = false;
             }
             log::info!(
                 "DeltaLake backfilling entries count: {} (elapsed time: {:?})",
-                backfilling_entries_queue.len(),
+                backfilling_entries.len(),
                 backfilling_started_at.elapsed()
             );
             Ok::<(), ReadError>(())
         })?;
         warn!("Backfilling thresholds won't be applied for any data the follows after the initially read batch.");
-        Ok(backfilling_entries_queue)
+
+        let is_pathway_output = has_pathway_meta_column || pathway_meta_column_added;
+        if is_pathway_output && !backfilling_entries.is_empty() {
+            let mut backfilling_entries_queue = VecDeque::new();
+            let mut prev_time = None;
+            backfilling_entries
+                .sort_by_key(|entry| entry.get(SPECIAL_FIELD_TIME).unwrap().clone().ok());
+            backfilling_entries_queue.push_back(BackfillingEntry::SourceEvent(
+                ReadResult::NewSource(ParquetMetadata::new(None).into()),
+            ));
+            for mut entry in backfilling_entries {
+                let current_time: Option<Value> =
+                    entry.get(SPECIAL_FIELD_TIME).unwrap().clone().ok().clone();
+                let is_new_block = prev_time.is_some() && current_time != prev_time;
+                if is_new_block {
+                    backfilling_entries_queue.push_back(BackfillingEntry::SourceEvent(
+                        ReadResult::FinishedSource {
+                            commit_allowed: true,
+                        },
+                    ));
+                    backfilling_entries_queue.push_back(BackfillingEntry::SourceEvent(
+                        ReadResult::NewSource(ParquetMetadata::new(None).into()),
+                    ));
+                }
+                if pathway_meta_column_added {
+                    // Pathway meta columns weren't requested by the user: they were added by us
+                    // artificially, to perform the data merging. Now they need to be removed from the
+                    // entry to correspond to what was requested.
+                    entry.remove(SPECIAL_FIELD_TIME);
+                }
+                backfilling_entries_queue.push_back(BackfillingEntry::Entry(entry));
+                prev_time = current_time;
+            }
+            backfilling_entries_queue.push_back(BackfillingEntry::SourceEvent(
+                ReadResult::FinishedSource {
+                    commit_allowed: true,
+                },
+            ));
+            if pathway_meta_column_added {
+                column_types.remove(SPECIAL_FIELD_TIME);
+            }
+            Ok(backfilling_entries_queue)
+        } else {
+            Ok(backfilling_entries
+                .into_iter()
+                .map(BackfillingEntry::Entry)
+                .collect())
+        }
     }
 
     fn scalar_value_for_queries(value: &Value) -> ScalarValue {
@@ -577,18 +680,70 @@ impl DeltaTableReader {
     fn get_reader_actions(
         table: &DeltaTable,
         base_path: &str,
+        mut history: Vec<DeltaTableCommitInfo>,
     ) -> Result<VecDeque<DeltaReaderAction>, ReadError> {
-        Ok(table
+        // Historical events without timestamps are useless for grouping parquet files
+        // into atomically processed versions, therefore there is a need to remove them
+        let original_history_len = history.len();
+        history.retain(|item| item.timestamp.is_some());
+        if history.len() != original_history_len {
+            warn!("Some of the historical entries don't have the timestamp, therefore some version updates may be merged into one. Original number of historical entries: {original_history_len}. Entries available after filtering: {}", history.len());
+        }
+        history.sort_by_key(|item| item.timestamp);
+
+        let mut actions_with_timestamp: Vec<_> = table
             .snapshot()?
             .file_actions()?
             .into_iter()
             .map(|action| {
-                DeltaReaderAction::new(
-                    DataEventType::Insert,
-                    Self::ensure_absolute_path_with_base(&action.path, base_path),
+                (
+                    DeltaReaderAction::new(
+                        DataEventType::Insert,
+                        Self::ensure_absolute_path_with_base(&action.path, base_path),
+                    ),
+                    action.modification_time,
                 )
             })
-            .collect())
+            .collect();
+        actions_with_timestamp.sort_by_key(|item| item.1);
+
+        let mut actions: VecDeque<DeltaReaderAction> =
+            VecDeque::with_capacity(actions_with_timestamp.len());
+        let mut n_total_versions = 0;
+        let mut current_commit_idx = 0;
+        for (action, timestamp) in actions_with_timestamp {
+            let mut is_new_block = false;
+            while current_commit_idx < history.len()
+                && timestamp
+                    > history[current_commit_idx]
+                        .timestamp
+                        .expect("events without timestamp have been filtered before")
+            {
+                // Every historical entry corresponds to a commit. If this action corresponds
+                // to a further historical entry, it means the start of the new atomically
+                // processed version block.
+                is_new_block = true;
+                current_commit_idx += 1;
+            }
+
+            if !actions.is_empty() && is_new_block {
+                actions
+                    .back_mut()
+                    .expect("actions are not empty")
+                    .set_last_in_version();
+                n_total_versions += 1;
+            }
+            actions.push_back(action);
+        }
+
+        // The last actions always terminates atomically processed batch
+        if let Some(last_action) = actions.back_mut() {
+            last_action.set_last_in_version();
+            n_total_versions += 1;
+        }
+
+        info!("The first read of Delta table at {base_path} uses {n_total_versions} versions and {} parquet files", actions.len());
+        Ok(actions)
     }
 
     fn ensure_absolute_path(&self, path: &str) -> String {
@@ -654,10 +809,13 @@ impl DeltaTableReader {
                     added_blocks.push_back(action);
                 }
 
-                self.last_fully_read_version = Some(self.current_version);
                 self.current_version = next_version;
                 self.rows_read_within_version = 0;
                 if data_changed {
+                    added_blocks
+                        .back_mut()
+                        .expect("if there is a data change, there should be at least one block")
+                        .set_last_in_version();
                     self.parquet_files_queue = added_blocks;
                 }
             }
@@ -665,48 +823,61 @@ impl DeltaTableReader {
         })
     }
 
-    fn read_next_row_native(&mut self, is_polling_enabled: bool) -> Result<ParquetRow, ReadError> {
-        loop {
-            if let Some(ref mut reader) = &mut self.reader {
-                match reader.next() {
-                    Some(Ok(row)) => return Ok(row),
-                    Some(Err(parquet_err)) => return Err(ReadError::Parquet(parquet_err)),
-                    None => self.reader = None,
-                };
-            } else {
-                if self.parquet_files_queue.is_empty() {
-                    self.upgrade_table_version(is_polling_enabled)?;
-                    if self.parquet_files_queue.is_empty() {
-                        return Err(ReadError::NoObjectsToRead);
-                    }
+    fn read_next_row_native(
+        &mut self,
+        is_polling_enabled: bool,
+    ) -> Result<ParquetReaderOutcome, ReadError> {
+        if let Some(ref mut reader) = &mut self.reader {
+            match reader.next() {
+                Some(Ok(row)) => return Ok(ParquetReaderOutcome::Row(row)),
+                Some(Err(parquet_err)) => return Err(ReadError::Parquet(parquet_err)),
+                None => {
+                    // The Pathway time advancement (e.g. commit) is only possible if it was the
+                    // last Parquet block within a version.
+                    let source_event = ReadResult::FinishedSource {
+                        commit_allowed: self
+                            .current_action
+                            .as_ref()
+                            .expect("current action must be set if there's a reader")
+                            .is_last_in_version,
+                    };
+                    self.reader = None;
+                    self.current_action = None;
+
+                    return Ok(ParquetReaderOutcome::SourceEvent(source_event));
                 }
-                let next_action = self.parquet_files_queue.pop_front().unwrap();
-                let local_object = self.object_downloader.download_object(&next_action.path)?;
-                self.current_event_type = next_action.action_type;
-                self.reader = Some(DeltaLakeParquetReader::try_from(local_object)?.into_iter());
+            };
+        }
+        if self.parquet_files_queue.is_empty() {
+            self.upgrade_table_version(is_polling_enabled)?;
+            if self.parquet_files_queue.is_empty() {
+                return Err(ReadError::NoObjectsToRead);
             }
         }
-    }
+        let next_action = self.parquet_files_queue.pop_front().unwrap();
+        let local_object = self.object_downloader.download_object(&next_action.path)?;
+        let new_block_metadata = ParquetMetadata::new(Some(next_action.path.clone()));
 
-    fn rows_in_file_count(path: &str) -> Result<i64, ReadError> {
-        let reader = DeltaLakeParquetReader::try_from(Path::new(path))?;
-        let metadata = reader.metadata();
-        let mut n_rows = 0;
-        for row_group in metadata.row_groups() {
-            n_rows += row_group.num_rows();
-        }
-        Ok(n_rows)
+        self.current_action = Some(next_action);
+        self.reader = Some(DeltaLakeParquetReader::try_from(local_object)?.into_iter());
+
+        let source_event = ReadResult::NewSource(new_block_metadata.into());
+        Ok(ParquetReaderOutcome::SourceEvent(source_event))
     }
 }
 
 impl Reader for DeltaTableReader {
     fn read(&mut self) -> Result<ReadResult, ReadError> {
-        let row_map = if let Some(row_map) = self.backfilling_entries_queue.pop_front() {
-            row_map
+        let row_map = if let Some(maybe_row_map) = self.backfilling_entries_queue.pop_front() {
+            match maybe_row_map {
+                BackfillingEntry::SourceEvent(event) => return Ok(event),
+                BackfillingEntry::Entry(entry) => entry,
+            }
         } else {
             let parquet_row =
                 match self.read_next_row_native(self.streaming_mode.is_polling_enabled()) {
-                    Ok(row) => row,
+                    Ok(ParquetReaderOutcome::Row(row)) => row,
+                    Ok(ParquetReaderOutcome::SourceEvent(event)) => return Ok(event),
                     Err(ReadError::NoObjectsToRead) => return Ok(ReadResult::Finished),
                     Err(other) => return Err(other),
                 };
@@ -715,13 +886,19 @@ impl Reader for DeltaTableReader {
 
         self.rows_read_within_version += 1;
         Ok(ReadResult::Data(
-            ReaderContext::from_diff(self.current_event_type, None, row_map),
+            ReaderContext::from_diff(
+                self.current_action
+                    .as_ref()
+                    // If there's no current action, backfilling thresholds were used. They only imply insertions
+                    .map_or(DataEventType::Insert, |action| action.action_type),
+                None,
+                row_map,
+            ),
             (
                 OffsetKey::Empty,
                 OffsetValue::DeltaTablePosition {
                     version: self.current_version,
                     rows_read_within_version: self.rows_read_within_version,
-                    last_fully_read_version: self.last_fully_read_version,
                 },
             ),
         ))
@@ -732,12 +909,7 @@ impl Reader for DeltaTableReader {
         // Then, the `seek` loads this checkpoint and ensures that no diffs
         // from the current version will be applied.
         let offset_value = frontier.get_offset(&OffsetKey::Empty);
-        let Some(OffsetValue::DeltaTablePosition {
-            version,
-            rows_read_within_version: n_rows_to_rewind,
-            last_fully_read_version,
-        }) = offset_value
-        else {
+        let Some(OffsetValue::DeltaTablePosition { version, .. }) = offset_value else {
             if offset_value.is_some() {
                 warn!("Incorrect type of offset value in DeltaLake frontier: {offset_value:?}");
             }
@@ -746,38 +918,11 @@ impl Reader for DeltaTableReader {
 
         self.reader = None;
         let runtime = create_async_tokio_runtime()?;
-        if let Some(last_fully_read_version) = last_fully_read_version {
-            // The offset is based on the diff between `last_fully_read_version` and `version`
-            self.current_version = *last_fully_read_version;
-            runtime.block_on(async { self.table.load_version(self.current_version).await })?;
-            self.upgrade_table_version(false)?;
-        } else {
-            // The offset is based on the full set of files present for `version`
-            self.current_version = *version;
-            runtime.block_on(async { self.table.load_version(self.current_version).await })?;
-            self.parquet_files_queue = Self::get_reader_actions(&self.table, &self.base_path)?;
-        }
 
-        self.rows_read_within_version = 0;
-        while !self.parquet_files_queue.is_empty() {
-            let next_block = self.parquet_files_queue.front().unwrap();
-            let block_size = Self::rows_in_file_count(&next_block.path)?;
-            if self.rows_read_within_version + block_size <= *n_rows_to_rewind {
-                info!(
-                    "Skipping parquet block with the size of {block_size} entries: {next_block:?}"
-                );
-                self.rows_read_within_version += block_size;
-                self.parquet_files_queue.pop_front();
-            } else {
-                break;
-            }
-        }
-
-        let rows_left_to_rewind = *n_rows_to_rewind - self.rows_read_within_version;
-        info!("Not quickly-rewindable entries count: {rows_left_to_rewind}");
-        for _ in 0..rows_left_to_rewind {
-            let _ = self.read_next_row_native(false)?;
-        }
+        // The last saved offset corresponds to the last version that has been read in full
+        self.current_version = *version;
+        runtime.block_on(async { self.table.load_version(self.current_version).await })?;
+        self.parquet_files_queue.clear();
 
         Ok(())
     }