+++ src/connectors/metadata/mod.rs
@@ -1,6 +1,7 @@
 pub mod file_like;
 pub mod iceberg;
 pub mod kafka;
+pub mod parquet;
 pub mod sqlite;
 
 #[allow(clippy::module_name_repetitions)]
@@ -13,6 +14,9 @@ pub use kafka::KafkaMetadata;
 pub use iceberg::IcebergMetadata;
 
 #[allow(clippy::module_name_repetitions)]
+pub use parquet::ParquetMetadata;
+
+#[allow(clippy::module_name_repetitions)]
 pub use sqlite::SQLiteMetadata;
 
 #[allow(clippy::module_name_repetitions)]
@@ -22,6 +26,7 @@ pub enum SourceMetadata {
     Kafka(KafkaMetadata),
     SQLite(SQLiteMetadata),
     Iceberg(IcebergMetadata),
+    Parquet(ParquetMetadata),
 }
 
 impl From<FileLikeMetadata> for SourceMetadata {
@@ -42,6 +47,12 @@ impl From<IcebergMetadata> for SourceMetadata {
     }
 }
 
+impl From<ParquetMetadata> for SourceMetadata {
+    fn from(impl_: ParquetMetadata) -> Self {
+        Self::Parquet(impl_)
+    }
+}
+
 impl From<SQLiteMetadata> for SourceMetadata {
     fn from(impl_: SQLiteMetadata) -> Self {
         Self::SQLite(impl_)
@@ -55,13 +66,14 @@ impl SourceMetadata {
             Self::Kafka(meta) => serde_json::to_value(meta),
             Self::SQLite(meta) => serde_json::to_value(meta),
             Self::Iceberg(meta) => serde_json::to_value(meta),
+            Self::Parquet(meta) => serde_json::to_value(meta),
         }
         .expect("Internal JSON serialization error")
     }
 
     pub fn commits_allowed_in_between(&self) -> bool {
         match self {
-            Self::FileLike(_) | Self::SQLite(_) | Self::Iceberg(_) => false,
+            Self::FileLike(_) | Self::SQLite(_) | Self::Iceberg(_) | Self::Parquet(_) => false,
             Self::Kafka(_) => true,
         }
     }