+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Changed
 - The `inactivity_detection` function is now a method for append only tables. It no longer relies on an event timestamp column but now uses table processing times to detect inactivity periods.
+- `pw.io.deltalake.read` now processes Delta table version updates atomically, applying all changes together in a single minibatch.
 
 ## [0.21.3] - 2025-04-16